Added helpers test class
